Add track icons to chat.
This feature is a bit buggy, and should cling to playback changes in the
room owner's player, rather than to everyone's, but I'm not sure how
that's implemented yet.  So for now, there'll be as many duplicates as
there are users in the room.

Also, I should figure out a way to ignore chat messages that happened
before you enter a party.