Wrap change:playingTrack callback.
The callback gets called twice, which is a little weird.

The first time, `newVal` is null, rather than the next track.  The
second time around, `newVal` contains the next track, leading me to
think that the player is (confusingly) stopping and playing after each
track.  While this literally makes sense, contextually, it's
awkward and cumbersome.